BUGFIX: totaltransactionscount now right behavior

+++ src/api/graphql/types/stats.rs
@@ -3,5 +3,5 @@ use async_graphql::SimpleObject;
 #[derive(SimpleObject)]
 pub struct Stats {
     #[graphql(name = "totalTransactionsCount")]
-    pub total_transactions_count: i32,
+    pub total_transactions_count: i64,
 }
